Calculate greyscale values for buckets
The MATLAB code now calculates the hex value for each bucket given the
fraction in each bucket per generation. Note that issue #17 needs to be
resolved for this to be totally accurate, but this will suffice until
death is implemented per the algorithm in  Brigatti et al which should
be soon.